% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{add_definition}
\alias{add_definition}
\title{Add a definition to an MCP object}
\usage{
add_definition(mcp, definition)
}
\arguments{
\item{mcp}{An MCP object}

\item{definition}{A definition object}
}
\value{
The MCP object with the definition added
}
\description{
Add a definition to an MCP object
}
