% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{build_param_info}
\alias{build_param_info}
\title{Build parameter information from @param and @type tags}
\usage{
build_param_info(param_tags, type_tags = list())
}
\arguments{
\item{param_tags}{List of @param tags}

\item{type_tags}{List of @type tags}
}
\value{
List of parameter information
}
\description{
Build parameter information from @param and @type tags
}
\keyword{internal}
