% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-conversion-utilities.R
\name{can_serialize}
\alias{can_serialize}
\title{Check if an R object can be safely serialized to JSON}
\usage{
can_serialize(x)
}
\arguments{
\item{x}{An R object to check}
}
\value{
TRUE if the object can be serialized, FALSE otherwise
}
\description{
Check if an R object can be safely serialized to JSON
}
