% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer-to-mcpr.R
\name{convert_ellmer_type_to_mcpr_property}
\alias{convert_ellmer_type_to_mcpr_property}
\title{Convert a single ellmer type to an mcpr property}
\usage{
convert_ellmer_type_to_mcpr_property(ellmer_type, prop_name)
}
\arguments{
\item{ellmer_type}{An ellmer Type object (TypeBasic, TypeArray, etc.)}

\item{prop_name}{The name of the property (for better error messages)}
}
\value{
An mcpr property object
}
\description{
Convert a single ellmer type to an mcpr property
}
\keyword{internal}
