% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer-to-mcpr.R
\name{convert_ellmer_types_to_mcpr}
\alias{convert_ellmer_types_to_mcpr}
\title{Convert ellmer TypeObject to mcpr schema}
\usage{
convert_ellmer_types_to_mcpr(type_object)
}
\arguments{
\item{type_object}{An ellmer TypeObject containing the function arguments}
}
\value{
An mcpr schema object
}
\description{
Convert ellmer TypeObject to mcpr schema
}
\keyword{internal}
