% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer.R
\name{create_ellmer_handler}
\alias{create_ellmer_handler}
\title{Create an ellmer handler function for an MCP tool}
\usage{
create_ellmer_handler(client, tool_name, input_schema)
}
\arguments{
\item{client}{The mcpr client}

\item{tool_name}{The name of the MCP tool}

\item{input_schema}{The tool's input schema}
}
\value{
A function that can be used as an ellmer tool handler
}
\description{
Create an ellmer handler function for an MCP tool
}
