% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpr-to-ellmer.R
\name{create_ellmer_types}
\alias{create_ellmer_types}
\title{Create ellmer type functions from MCP schema properties}
\usage{
create_ellmer_types(schema)
}
\arguments{
\item{schema}{The MCP schema object}
}
\value{
A list of ellmer type objects for each property
}
\description{
Create ellmer type functions from MCP schema properties
}
\keyword{internal}
