% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonrpc.R
\name{create_error}
\alias{create_error}
\title{Create a JSON-RPC 2.0 error response}
\usage{
create_error(code, message, data = NULL, id = NULL)
}
\arguments{
\item{code}{Error code}

\item{message}{Error message}

\item{data}{Optional error data}

\item{id}{Request identifier}
}
\value{
A structured JSON-RPC 2.0 error response
}
\description{
Create a JSON-RPC 2.0 error response
}
\keyword{internal}
