% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{daemon_sockets}
\alias{daemon_sockets}
\title{Daemon Socket Registry}
\description{
Named list of nanonext sockets for MCPR-owned secondary sessions.
Keys are internal secondary-session IDs, values are nanonext socket objects.
}
