% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{decorators}
\alias{decorators}
\title{Parse MCP Decorators from R Source}
\description{
Functions for parsing roxygen2-style decorators (@mcp_tool, @mcp_resource, @mcp_prompt)
from R source files. This enables a decorator-based approach similar to plumber.
}
