% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-conversion-serializers.R
\docType{data}
\name{.mcp_custom_serializers}
\alias{.mcp_custom_serializers}
\title{Global registry for custom serializers}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.mcp_custom_serializers
}
\description{
Global registry for custom serializers
}
\keyword{internal}
