% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer-to-mcpr.R
\name{ellmer_to_mcpr_tool}
\alias{ellmer_to_mcpr_tool}
\title{Convert ellmer tools to mcpr tools}
\usage{
ellmer_to_mcpr_tool(ellmer_tool)
}
\arguments{
\item{ellmer_tool}{An ellmer ToolDef object created with \code{ellmer::tool()}}
}
\value{
An mcpr tool object compatible with \code{new_tool()}
}
\description{
This function converts tools from the ellmer package to a format compatible
with the mcpr package. It takes an ellmer ToolDef object and creates an
mcpr tool object with the appropriate input schema and handler function.
}
\examples{
\dontrun{
# Create an ellmer tool
ellmer_rnorm <- ellmer::tool(
  rnorm,
  "Generate random normal numbers",
  n = ellmer::type_integer("Number of observations"),
  mean = ellmer::type_number("Mean value", required = FALSE),
  sd = ellmer::type_number("Standard deviation", required = FALSE)
)

# Convert to mcpr format
mcpr_tool <- ellmer_to_mcpr_tool(ellmer_rnorm)

# Add to an mcpr server
server <- new_server("MyServer", "Test server", "1.0.0")
add_capability(server, mcpr_tool)
}

}
\seealso{
\code{\link[=new_tool]{new_tool()}}, \code{\link[ellmer:tool]{ellmer::tool()}}, \code{\link[=add_capability]{add_capability()}}
}
