% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{format_prompt_registrations}
\alias{format_prompt_registrations}
\title{Format prompt registrations for protocol server}
\usage{
format_prompt_registrations(prompts)
}
\arguments{
\item{prompts}{List of prompt definitions}
}
\value{
Character vector of R code lines
}
\description{
Format prompt registrations for protocol server
}
\keyword{internal}
