% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{format_prompts_definition}
\alias{format_prompts_definition}
\title{Format Prompts Definition for R Code}
\usage{
format_prompts_definition(prompts)
}
\arguments{
\item{prompts}{List of prompt definitions}
}
\value{
Formatted R code string
}
\description{
Format Prompts Definition for R Code
}
\keyword{internal}
