% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{format_resource_registrations}
\alias{format_resource_registrations}
\title{Format resource registrations for protocol server}
\usage{
format_resource_registrations(resources)
}
\arguments{
\item{resources}{List of resource definitions}
}
\value{
Character vector of R code lines
}
\description{
Format resource registrations for protocol server
}
\keyword{internal}
