% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-server-formatters.R
\name{format_resources_definition}
\alias{format_resources_definition}
\title{Format Resources Definition for R Code}
\usage{
format_resources_definition(resources)
}
\arguments{
\item{resources}{List of resource definitions}
}
\value{
Formatted R code string
}
\description{
Format Resources Definition for R Code
}
\keyword{internal}
