% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{format_tool_registrations}
\alias{format_tool_registrations}
\title{Format tool registrations for protocol server}
\usage{
format_tool_registrations(tools)
}
\arguments{
\item{tools}{List of tool definitions}
}
\value{
Character vector of R code lines
}
\description{
Format tool registrations for protocol server
}
\keyword{internal}
