% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-server-formatters.R
\name{format_tools_definition}
\alias{format_tools_definition}
\title{Format Tools Definition for R Code}
\usage{
format_tools_definition(tools)
}
\arguments{
\item{tools}{List of tool definitions}
}
\value{
Formatted R code string
}
\description{
Format Tools Definition for R Code
}
\keyword{internal}
