% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonrpc.R
\name{from_json}
\alias{from_json}
\title{Parse JSON to an R object}
\usage{
from_json(json, ...)
}
\arguments{
\item{json}{JSON string to parse}

\item{...}{Additional arguments passed to jsonlite::fromJSON}
}
\value{
R object
}
\description{
Parse JSON to an R object
}
\keyword{internal}
