% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-server-files.R
\name{generate_gitignore}
\alias{generate_gitignore}
\title{Generate .gitignore}
\usage{
generate_gitignore(server_dir)
}
\arguments{
\item{server_dir}{Server directory path}
}
\description{
Generate .gitignore
}
\keyword{internal}
