% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{generate_handler_code}
\alias{generate_handler_code}
\title{Generate handler code that calls the original function}
\usage{
generate_handler_code(tool)
}
\arguments{
\item{tool}{Tool information}
}
\value{
Character vector of handler code lines
}
\description{
Generate handler code that calls the original function
}
\keyword{internal}
