% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{generate_package_json}
\alias{generate_package_json}
\title{Generate package.json}
\usage{
generate_package_json(server_dir, vars)
}
\arguments{
\item{server_dir}{Server directory path}

\item{vars}{Template variables}
}
\description{
Generate package.json
}
\keyword{internal}
