% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{generate_property_code}
\alias{generate_property_code}
\title{Generate property code for a parameter}
\usage{
generate_property_code(param)
}
\arguments{
\item{param}{Parameter information}
}
\value{
Character string with property code
}
\description{
Generate property code for a parameter
}
\keyword{internal}
