% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{generate_protocol_server}
\alias{generate_protocol_server}
\title{Generate protocol-enhanced R server}
\usage{
generate_protocol_server(
  server_dir,
  vars,
  tools,
  resources,
  prompts,
  source_files
)
}
\arguments{
\item{server_dir}{Server directory path}

\item{vars}{Template variables}

\item{tools}{Tool definitions}

\item{resources}{Resource definitions}

\item{prompts}{Prompt definitions}

\item{source_files}{Original source files to copy}
}
\description{
Generate protocol-enhanced R server
}
\keyword{internal}
