% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-server-files.R
\name{generate_r_server}
\alias{generate_r_server}
\title{Generate R Server}
\usage{
generate_r_server(server_dir, vars, tools, resources, prompts, template)
}
\arguments{
\item{server_dir}{Server directory path}

\item{vars}{Template variables}

\item{tools}{Tool definitions}

\item{resources}{Resource definitions}

\item{prompts}{Prompt definitions}

\item{template}{Template type}
}
\description{
Generate R Server
}
\keyword{internal}
