% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{generate_readme}
\alias{generate_readme}
\title{Generate README.md}
\usage{
generate_readme(server_dir, vars)
}
\arguments{
\item{server_dir}{Server directory path}

\item{vars}{Template variables}
}
\description{
Generate README.md
}
\keyword{internal}
