% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{generate_test_script}
\alias{generate_test_script}
\title{Generate Test Script}
\usage{
generate_test_script(server_dir, vars)
}
\arguments{
\item{server_dir}{Server directory path}

\item{vars}{Template variables}
}
\description{
Generate Test Script
}
\keyword{internal}
