% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{generate_tool_code}
\alias{generate_tool_code}
\title{Generate R code for a single tool}
\usage{
generate_tool_code(tool, index)
}
\arguments{
\item{tool}{Tool information list}

\item{index}{Tool index for naming}
}
\value{
Character vector of R code lines
}
\description{
Generate R code for a single tool
}
\keyword{internal}
