% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_server.R
\name{generate_wrapper}
\alias{generate_wrapper}
\title{Generate Node.js Wrapper}
\usage{
generate_wrapper(server_dir, vars)
}
\arguments{
\item{server_dir}{Server directory path}

\item{vars}{Template variables}
}
\description{
Generate Node.js Wrapper
}
\keyword{internal}
