% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-middleware.R
\name{get_request_audience}
\alias{get_request_audience}
\title{Get request audience for token validation}
\usage{
get_request_audience(req)
}
\arguments{
\item{req}{HTTP request object}
}
\value{
Audience string
}
\description{
Get request audience for token validation
}
