% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{input_schema}
\alias{input_schema}
\title{Create a new input schema}
\usage{
input_schema(properties, additional_properties = FALSE)
}
\arguments{
\item{properties}{List of property definitions}

\item{additional_properties}{Logical indicating if additional properties are allowed}
}
\value{
A list representing a JSON Schema object
}
\description{
Create a new input schema
}
\examples{
schema <- input_schema(
  properties = list(
    name = prop_string("User name", "The name of the user", required = TRUE),
    age = prop_number("User age", "The age of the user in years", minimum = 0)
  )
)
}
