% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-utils.R
\name{is_oauth_endpoint}
\alias{is_oauth_endpoint}
\title{Check if a path is an OAuth endpoint}
\usage{
is_oauth_endpoint(path)
}
\arguments{
\item{path}{Request path}
}
\value{
Logical indicating if path is OAuth-related
}
\description{
Check if a path is an OAuth endpoint
}
