% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonrpc.R
\docType{data}
\name{JSONRPC_PARSE_ERROR}
\alias{JSONRPC_PARSE_ERROR}
\title{JSON-RPC 2.0 Standard Error Codes}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
JSONRPC_PARSE_ERROR
}
\description{
JSON-RPC 2.0 Standard Error Codes
}
\keyword{internal}
