% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_utils.R
\name{load_server_config}
\alias{load_server_config}
\title{Load Server Configuration from File}
\usage{
load_server_config(config_file)
}
\arguments{
\item{config_file}{Path to configuration file}
}
\value{
Configuration list
}
\description{
Loads server configuration from a YAML or JSON file
}
