% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-conversion-utilities.R
\name{mcp_serialize}
\alias{mcp_serialize}
\title{Serialize R object to JSON string for MCP}
\usage{
mcp_serialize(x, pretty = FALSE, auto_unbox = TRUE)
}
\arguments{
\item{x}{An R object to serialize}

\item{pretty}{Whether to pretty-print the JSON}

\item{auto_unbox}{Whether to automatically unbox single-element vectors}
}
\value{
A JSON string representation of the R object
}
\description{
Serialize R object to JSON string for MCP
}
\examples{
mcp_serialize(list(result = 42, message = "success"))
}
