% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpr-to-ellmer.R
\name{mcpr_to_ellmer_tools}
\alias{mcpr_to_ellmer_tools}
\title{Convert MCPR tools to ellmer tools}
\usage{
mcpr_to_ellmer_tools(client)
}
\arguments{
\item{client}{An mcpr client object}
}
\value{
A list of ellmer-compatible tool functions
}
\description{
This function converts tools from an MCPR client to a format compatible
with the ellmer package. It retrieves all available tools from the client
and creates wrapper functions that call these tools through the MCPR protocol.
}
\examples{
\dontrun{
# Create an MCPR client
client <- new_client_io("path/to/server")

# Convert its tools to ellmer format
ellmer_tools <- mcpr_to_ellmer_tools(client)

# Use with ellmer
chat <- ellmer::chat_claude()
chat$set_tools(ellmer_tools)
}
}
\keyword{internal}
