% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{new_client}
\alias{new_client}
\title{Create a new mcp client}
\usage{
new_client(command, args = character(), name = command, version = "1.0.0")
}
\arguments{
\item{command}{The command to run}

\item{args}{Arguments to pass to the command}

\item{name}{The name of the client}

\item{version}{The version of the client}
}
\value{
A new mcp client
}
\description{
Create a new mcp client
}
