% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{new_prompt}
\alias{new_prompt}
\title{Create a new prompt}
\usage{
new_prompt(name, description, arguments = list())
}
\arguments{
\item{name}{Name of the prompt}

\item{description}{Description of the prompt}

\item{arguments}{List of arguments for the prompt}
}
\value{
A new prompt definition
}
\description{
Create a new prompt
}
\examples{
prompt <- new_prompt(
  name = "My Prompt",
  description = "This is a description",
  arguments = list(
    input1 = list(
      type = "string",
      description = "Input 1"
    ),
    input2 = list(
      type = "number",
      description = "Input 2"
    )
  )
)
}
