% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{new_property}
\alias{new_property}
\title{Create a new property}
\usage{
new_property(type, title, description, required = FALSE, ...)
}
\arguments{
\item{type}{Type of the property}

\item{title}{Short title for the property}

\item{description}{Longer description of the property}

\item{required}{Whether the property is required}

\item{...}{Additional attributes for the property}
}
\value{
A property object with the specified attributes
}
\description{
Create a new property
}
\keyword{internal}
