% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{new_resource}
\alias{new_resource}
\title{Create a new resource}
\usage{
new_resource(name, description, uri, mime_type = NULL, handler)
}
\arguments{
\item{name}{Name of the resource}

\item{description}{Description of the resource}

\item{uri}{URI of the resource}

\item{mime_type}{MIME type of the resource (optional)}

\item{handler}{Function to handle the resource request}
}
\value{
A new resource capability
}
\description{
Create a new resource
}
\examples{
resource <- new_resource(
  name = "My Resource",
  description = "This is a description",
  uri = "https://example.com/resource",
  mime_type = "text/plain",
  handler = function(params) {
    # Process the resource request
    return(list(content = "Resource content"))
  }
)
}
