% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-server.R
\name{oauth_authorization_endpoint}
\alias{oauth_authorization_endpoint}
\title{Authorization endpoint handler}
\usage{
oauth_authorization_endpoint(oauth_server)
}
\arguments{
\item{oauth_server}{OAuth server object}
}
\value{
Function for handling authorization requests
}
\description{
Authorization endpoint handler
}
