% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_cleanup_expired}
\alias{oauth_cleanup_expired}
\title{Clean expired tokens and codes}
\usage{
oauth_cleanup_expired(storage)
}
\arguments{
\item{storage}{Storage object}
}
\value{
Number of items cleaned
}
\description{
Clean expired tokens and codes
}
