% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_consume_authorization_code}
\alias{oauth_consume_authorization_code}
\title{Retrieve and consume authorization code}
\usage{
oauth_consume_authorization_code(storage, code)
}
\arguments{
\item{storage}{Storage object}

\item{code}{Authorization code}
}
\value{
Code information or NULL if not found/expired/used
}
\description{
Retrieve and consume authorization code
}
