% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth_create_test_client}
\alias{oauth_create_test_client}
\title{Generate OAuth client for testing}
\usage{
oauth_create_test_client(
  oauth_server,
  client_name = "Test MCP Client",
  redirect_uri = "http://localhost:3000/callback"
)
}
\arguments{
\item{oauth_server}{OAuth server object}

\item{client_name}{Name for the test client}

\item{redirect_uri}{Redirect URI for the client}
}
\value{
List with client credentials
}
\description{
Generate OAuth client for testing
}
