% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-utils.R
\name{oauth_error_response}
\alias{oauth_error_response}
\title{Parse OAuth error parameters}
\usage{
oauth_error_response(error, error_description = NULL, error_uri = NULL)
}
\arguments{
\item{error}{Error code}

\item{error_description}{Error description}

\item{error_uri}{Optional error URI}
}
\value{
List with OAuth error response
}
\description{
Parse OAuth error parameters
}
