% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth_exchange_code}
\alias{oauth_exchange_code}
\title{Exchange authorization code for access token}
\usage{
oauth_exchange_code(
  oauth_server,
  client_id,
  client_secret = NULL,
  code,
  redirect_uri,
  code_verifier
)
}
\arguments{
\item{oauth_server}{OAuth server object}

\item{client_id}{OAuth client ID}

\item{client_secret}{OAuth client secret}

\item{code}{Authorization code}

\item{redirect_uri}{Redirect URI}

\item{code_verifier}{PKCE code verifier}
}
\value{
List with token response
}
\description{
Exchange authorization code for access token
}
