% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_file_storage}
\alias{oauth_file_storage}
\title{Create file-based OAuth storage}
\usage{
oauth_file_storage(path)
}
\arguments{
\item{path}{Path to storage file}
}
\value{
File storage object
}
\description{
Create file-based OAuth storage
}
