% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth_generate_auth_url}
\alias{oauth_generate_auth_url}
\title{Generate OAuth authorization URL with PKCE}
\usage{
oauth_generate_auth_url(
  oauth_server,
  client_id,
  redirect_uri,
  scope = NULL,
  state = NULL
)
}
\arguments{
\item{oauth_server}{OAuth server object}

\item{client_id}{OAuth client ID}

\item{redirect_uri}{Redirect URI}

\item{scope}{Requested scopes}

\item{state}{Optional state parameter}
}
\value{
List with authorization URL and PKCE parameters
}
\description{
Generate OAuth authorization URL with PKCE
}
