% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-utils.R
\name{oauth_generate_authorization_code}
\alias{oauth_generate_authorization_code}
\title{Generate authorization code}
\usage{
oauth_generate_authorization_code(length = 32)
}
\arguments{
\item{length}{Code length (default: 32)}
}
\value{
Authorization code string
}
\description{
Generate authorization code
}
