% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-utils.R
\name{oauth_generate_keypair}
\alias{oauth_generate_keypair}
\title{Generate OAuth keypair for JWT signing}
\usage{
oauth_generate_keypair()
}
\value{
List with private key, public key, and public JWKS
}
\description{
Generate OAuth keypair for JWT signing
}
