% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_get_access_token}
\alias{oauth_get_access_token}
\title{Get access token information}
\usage{
oauth_get_access_token(storage, token_id)
}
\arguments{
\item{storage}{Storage object}

\item{token_id}{Token identifier}
}
\value{
Token information or NULL if not found
}
\description{
Get access token information
}
