% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-middleware.R
\name{oauth_get_auth_context}
\alias{oauth_get_auth_context}
\title{Create authentication context from request}
\usage{
oauth_get_auth_context(req)
}
\arguments{
\item{req}{HTTP request object}
}
\value{
Authentication context list or NULL
}
\description{
Create authentication context from request
}
