% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_get_client}
\alias{oauth_get_client}
\title{Retrieve OAuth client information}
\usage{
oauth_get_client(storage, client_id)
}
\arguments{
\item{storage}{Storage object}

\item{client_id}{Client identifier}
}
\value{
Client information or NULL if not found
}
\description{
Retrieve OAuth client information
}
