% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-middleware.R
\name{oauth_get_user_scopes}
\alias{oauth_get_user_scopes}
\title{Get user scopes from authenticated request}
\usage{
oauth_get_user_scopes(req)
}
\arguments{
\item{req}{HTTP request object}
}
\value{
Character vector of scopes or NULL
}
\description{
Get user scopes from authenticated request
}
