% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth_introspection_endpoint}
\alias{oauth_introspection_endpoint}
\title{OAuth token introspection endpoint (RFC 7662)}
\usage{
oauth_introspection_endpoint(oauth_server)
}
\arguments{
\item{oauth_server}{OAuth server object}
}
\value{
Function for handling introspection requests
}
\description{
OAuth token introspection endpoint (RFC 7662)
}
