% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-middleware.R
\name{oauth_is_authenticated}
\alias{oauth_is_authenticated}
\title{Check if request is authenticated}
\usage{
oauth_is_authenticated(req)
}
\arguments{
\item{req}{HTTP request object}
}
\value{
Logical indicating if request has valid authentication
}
\description{
Check if request is authenticated
}
