% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-server.R
\name{oauth_jwks_endpoint}
\alias{oauth_jwks_endpoint}
\title{Get JWKS endpoint handler}
\usage{
oauth_jwks_endpoint(oauth_server)
}
\arguments{
\item{oauth_server}{OAuth server object}
}
\value{
Function for handling JWKS requests
}
\description{
Get JWKS endpoint handler
}
