% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_memory_storage}
\alias{oauth_memory_storage}
\title{Create in-memory OAuth storage}
\usage{
oauth_memory_storage()
}
\value{
Memory storage object
}
\description{
Create in-memory OAuth storage
}
