% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-server.R
\name{oauth_register_client_endpoint}
\alias{oauth_register_client_endpoint}
\title{Dynamic client registration endpoint handler (RFC 7591)}
\usage{
oauth_register_client_endpoint(oauth_server)
}
\arguments{
\item{oauth_server}{OAuth server object}
}
\value{
Function for handling client registration
}
\description{
Dynamic client registration endpoint handler (RFC 7591)
}
