% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-middleware.R
\name{oauth_require_scopes}
\alias{oauth_require_scopes}
\title{Require specific OAuth scopes}
\usage{
oauth_require_scopes(required_scopes)
}
\arguments{
\item{required_scopes}{Vector of required scope strings}
}
\value{
Middleware function that checks for required scopes
}
\description{
Require specific OAuth scopes
}
