% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-server.R
\name{oauth_resource_metadata_endpoint}
\alias{oauth_resource_metadata_endpoint}
\title{Get OAuth protected resource metadata endpoint handler}
\usage{
oauth_resource_metadata_endpoint(oauth_server)
}
\arguments{
\item{oauth_server}{OAuth server object}
}
\value{
Function for handling resource metadata requests
}
\description{
Get OAuth protected resource metadata endpoint handler
}
