% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth_revocation_endpoint}
\alias{oauth_revocation_endpoint}
\title{OAuth token revocation endpoint (RFC 7009)}
\usage{
oauth_revocation_endpoint(oauth_server)
}
\arguments{
\item{oauth_server}{OAuth server object}
}
\value{
Function for handling revocation requests
}
\description{
OAuth token revocation endpoint (RFC 7009)
}
