% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_revoke_access_token}
\alias{oauth_revoke_access_token}
\title{Revoke access token}
\usage{
oauth_revoke_access_token(storage, token_id)
}
\arguments{
\item{storage}{Storage object}

\item{token_id}{Token identifier}
}
\value{
Logical indicating success
}
\description{
Revoke access token
}
