% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-middleware.R
\name{oauth_send_auth_error}
\alias{oauth_send_auth_error}
\title{Send OAuth authentication error response}
\usage{
oauth_send_auth_error(res, error, error_description = NULL, error_uri = NULL)
}
\arguments{
\item{res}{HTTP response object}

\item{error}{OAuth error code}

\item{error_description}{Optional error description}

\item{error_uri}{Optional error documentation URI}
}
\description{
Send OAuth authentication error response
}
