% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_storage_stats}
\alias{oauth_storage_stats}
\title{Get storage statistics}
\usage{
oauth_storage_stats(storage)
}
\arguments{
\item{storage}{Storage object}
}
\value{
List with storage statistics
}
\description{
Get storage statistics
}
