% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_store_access_token}
\alias{oauth_store_access_token}
\title{Store access token}
\usage{
oauth_store_access_token(storage, token_id, token_info)
}
\arguments{
\item{storage}{Storage object}

\item{token_id}{Token identifier (jti claim)}

\item{token_info}{Token information}
}
\description{
Store access token
}
