% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_store_authorization_code}
\alias{oauth_store_authorization_code}
\title{Store authorization code}
\usage{
oauth_store_authorization_code(storage, code, code_info, expires_in = 600)
}
\arguments{
\item{storage}{Storage object}

\item{code}{Authorization code}

\item{code_info}{Code information (client_id, redirect_uri, etc.)}

\item{expires_in}{Expiration time in seconds (default: 600)}
}
\description{
Store authorization code
}
