% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-storage.R
\name{oauth_store_client}
\alias{oauth_store_client}
\title{Store OAuth client information}
\usage{
oauth_store_client(storage, client_id, client_info)
}
\arguments{
\item{storage}{Storage object}

\item{client_id}{Client identifier}

\item{client_info}{Client information list}
}
\description{
Store OAuth client information
}
