% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-server.R
\name{oauth_token_endpoint}
\alias{oauth_token_endpoint}
\title{Token endpoint handler}
\usage{
oauth_token_endpoint(oauth_server)
}
\arguments{
\item{oauth_server}{OAuth server object}
}
\value{
Function for handling token requests
}
\description{
Token endpoint handler
}
