% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-middleware.R
\name{oauth_validate_bearer_token}
\alias{oauth_validate_bearer_token}
\title{Validate OAuth bearer token}
\usage{
oauth_validate_bearer_token(oauth_server, token, req)
}
\arguments{
\item{oauth_server}{OAuth server object}

\item{token}{Bearer token string}

\item{req}{HTTP request object}
}
\value{
List with validation results
}
\description{
Validate OAuth bearer token
}
