% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators-parse.R
\name{parse_mcp_decorators}
\alias{parse_mcp_decorators}
\title{Parse decorators from a source file}
\usage{
parse_mcp_decorators(file)
}
\arguments{
\item{file}{Path to R source file}
}
\value{
A list of parsed elements with their decorators and function definitions
}
\description{
Parse decorators from a source file
}
