% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonrpc.R
\name{process_batch}
\alias{process_batch}
\title{Process batch JSON-RPC 2.0 requests}
\usage{
process_batch(batch_requests, mcp)
}
\arguments{
\item{batch_requests}{List of request objects}

\item{mcp}{MCP server object}
}
\value{
List of response objects
}
\description{
Process batch JSON-RPC 2.0 requests
}
