% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{prompts_list}
\alias{prompts_list}
\title{List all available prompts}
\usage{
prompts_list(mcp)
}
\arguments{
\item{mcp}{A server object}
}
\value{
A list containing all available prompts
}
\description{
List all available prompts
}
