% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{prop_boolean}
\alias{prop_boolean}
\title{Create a boolean property definition}
\usage{
prop_boolean(title, description, required = FALSE)
}
\arguments{
\item{title}{Short title for the property}

\item{description}{Longer description of the property}

\item{required}{Whether the property is required}
}
\value{
A boolean property object
}
\description{
Create a boolean property definition
}
\examples{
active_prop <- prop_boolean(
  "Active status",
  "Whether the user account is active",
  required = TRUE
)
}
