% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{properties}
\alias{properties}
\title{Create a new properties list}
\usage{
properties(...)
}
\arguments{
\item{...}{Property objects}
}
\value{
A list of property objects
}
\description{
Create a new properties list
}
\examples{
properties <- properties(
  property_string("Name", "The name of the user", required = TRUE),
  property_number("Age", "The age of the user in years", minimum = 0)
)
}
