% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{property_array}
\alias{property_array}
\title{Create an array property definition}
\usage{
property_array(
  title,
  description,
  items,
  required = FALSE,
  min_items = NULL,
  max_items = NULL,
  unique_items = FALSE
)
}
\arguments{
\item{title}{Short title for the property}

\item{description}{Longer description of the property}

\item{items}{Schema for the items in the array}

\item{required}{Whether the property is required}

\item{min_items}{Optional minimum number of items}

\item{max_items}{Optional maximum number of items}

\item{unique_items}{Logical indicating if items must be unique}
}
\value{
An array property object
}
\description{
Create an array property definition
}
\examples{
tags_prop <- property_array(
  "Tags",
  "List of tags for the user",
  items = property_string("Tag", "A user tag"),
  unique_items = TRUE
)
}
