% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{property_number}
\alias{property_number}
\title{Create a number property definition}
\usage{
property_number(
  title,
  description,
  required = FALSE,
  minimum = NULL,
  maximum = NULL,
  exclusive_minimum = NULL,
  exclusive_maximum = NULL,
  multiple_of = NULL,
  integer = FALSE
)
}
\arguments{
\item{title}{Short title for the property}

\item{description}{Longer description of the property}

\item{required}{Whether the property is required}

\item{minimum}{Optional minimum value}

\item{maximum}{Optional maximum value}

\item{exclusive_minimum}{Whether minimum is exclusive}

\item{exclusive_maximum}{Whether maximum is exclusive}

\item{multiple_of}{Optional value the number must be a multiple of}

\item{integer}{Whether the number should be an integer}
}
\value{
A number property object
}
\description{
Create a number property definition
}
\examples{
age_prop <- property_number(
  "User age",
  "The age of the user in years",
  required = TRUE,
  minimum = 0,
  integer = TRUE
)
}
