% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{property_object}
\alias{property_object}
\title{Create an object property definition}
\usage{
property_object(
  title,
  description,
  properties,
  required = FALSE,
  additional_properties = FALSE
)
}
\arguments{
\item{title}{Short title for the property}

\item{description}{Longer description of the property}

\item{properties}{List of property definitions for this object}

\item{required}{Whether the property is required}

\item{additional_properties}{Logical indicating if additional properties are allowed}
}
\value{
An object property object
}
\description{
Create an object property definition
}
\examples{
address_prop <- property_object(
  "Address",
  "User's address information",
  properties = list(
    street = property_string("Street", "Street address", required = TRUE),
    city = property_string("City", "City name", required = TRUE),
    country = property_string("Country", "Country name")
  )
)
}
