% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{property_string}
\alias{property_string}
\title{Create a string property definition}
\usage{
property_string(
  title,
  description,
  required = FALSE,
  enum = NULL,
  pattern = NULL,
  min_length = NULL,
  max_length = NULL,
  format = NULL
)
}
\arguments{
\item{title}{Short title for the property}

\item{description}{Longer description of the property}

\item{required}{Whether the property is required}

\item{enum}{Optional character vector of allowed values}

\item{pattern}{Optional regex pattern the string must match}

\item{min_length}{Optional minimum length}

\item{max_length}{Optional maximum length}

\item{format}{Optional format constraint}
}
\value{
A string property object
}
\description{
Create a string property definition
}
\examples{
name_prop <- property_string(
  "User name",
  "The name of the user",
  required = TRUE,
  min_length = 2,
  max_length = 50
)
}
