% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{read}
\alias{read}
\title{Read a JSON-RPC response from a client provider}
\usage{
read(x, timeout = 60 * 1000)
}
\arguments{
\item{x}{A client provider}

\item{timeout}{Timeout in milliseconds for reading the response}
}
\value{
The response
}
\description{
Read a JSON-RPC response from a client provider
}
