% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpr-to-ellmer.R
\name{register_mcpr_tools}
\alias{register_mcpr_tools}
\title{Register MCPR tools with an ellmer chat}
\usage{
register_mcpr_tools(chat, client)
}
\arguments{
\item{chat}{An ellmer chat object}

\item{client}{An mcpr client object}
}
\value{
The chat object (invisibly)
}
\description{
This function registers tools from an MCPR client with an ellmer chat instance.
}
