% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{resources_list}
\alias{resources_list}
\title{List all available resources}
\usage{
resources_list(mcp)
}
\arguments{
\item{mcp}{A server object}
}
\value{
A list containing all available resources
}
\description{
List all available resources
}
