% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{resources_read}
\alias{resources_read}
\title{Read a resource with the given parameters}
\usage{
resources_read(mcp, params, id = NULL)
}
\arguments{
\item{mcp}{A server object}

\item{params}{Parameters for the resource read}

\item{id}{Optional request ID for response tracking}
}
\value{
A response object with the resource read results or an error
}
\description{
Read a resource with the given parameters
}
