% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roclet_process.roclet_mcp}
\alias{roclet_process.roclet_mcp}
\title{Process blocks for MCP roclet}
\usage{
\method{roclet_process}{roclet_mcp}(x, blocks, env, base_path)
}
\arguments{
\item{x}{MCP roclet object}

\item{blocks}{List of roxy_block objects}

\item{env}{Environment}

\item{base_path}{Base path}
}
\value{
List of processed MCP tools
}
\description{
Process blocks for MCP roclet
}
