% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roxy_tag_parse.roxy_tag_type}
\alias{roxy_tag_parse.roxy_tag_type}
\title{Parse @type tag}
\usage{
\method{roxy_tag_parse}{roxy_tag_type}(x)
}
\arguments{
\item{x}{A roxy_tag object}
}
\description{
Parses @type tags to extract parameter type information.
Format: @type param_name type enum_values
}
