% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roxy_tag_rd.roxy_tag_mcp}
\alias{roxy_tag_rd.roxy_tag_mcp}
\title{Roxygen2 tag for @mcp
This function is called by Roxygen2 to generate documentation for the @mcp tag}
\usage{
\method{roxy_tag_rd}{roxy_tag_mcp}(x, base_path, env)
}
\arguments{
\item{x}{Roxygen2 tag object}

\item{base_path}{Base path for the package}

\item{env}{Environment}
}
\value{
NULL (invisible)
}
\description{
Roxygen2 tag for @mcp
This function is called by Roxygen2 to generate documentation for the @mcp tag
}
