% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roxy_tag_rd.roxy_tag_type}
\alias{roxy_tag_rd.roxy_tag_type}
\title{Roxygen2 tag handler for @type
This function is called by Roxygen2 to generate documentation for the @type}
\usage{
\method{roxy_tag_rd}{roxy_tag_type}(x, base_path, env)
}
\arguments{
\item{x}{Roxygen2 tag object}

\item{base_path}{Base path for the package}

\item{env}{Environment}
}
\value{
NULL (invisible)
}
\description{
Roxygen2 tag handler for @type
This function is called by Roxygen2 to generate documentation for the @type
}
