% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-schema.R
\name{schema}
\alias{schema}
\title{Create a new input schema}
\usage{
schema(properties, type = "object", additional_properties = FALSE)
}
\arguments{
\item{properties}{List of property definitions}

\item{type}{Type of the schema}

\item{additional_properties}{Logical indicating if additional properties are allowed}
}
\value{
A list representing a JSON Schema object
}
\description{
Create a new input schema
}
\examples{
schema <- schema(
  properties = properties(
    name = property_string("User name", "The name of the user", required = TRUE),
    age = property_number("User age", "The age of the user in years", minimum = 0)
  )
)
}
