% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{serve_io}
\alias{serve_io}
\title{Serve an MCP server using stdin/stdout}
\usage{
serve_io(mcp)
}
\arguments{
\item{mcp}{An MCP server object created with new_server()}
}
\value{
Nothing, runs indefinitely in normal mode, or the response in test mode
}
\description{
Serve an MCP server using stdin/stdout
}
