% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{serve_std_io}
\alias{serve_std_io}
\title{Serve an MCP server using stdin/stdout}
\usage{
serve_std_io(mcp)
}
\arguments{
\item{mcp}{An MCP server object}

\item{input_con}{Connection to read from (defaults to stdin())}

\item{output_con}{Connection to write to (defaults to stdout())}
}
\value{
Nothing, runs indefinitely in normal mode, or the response in test mode
}
\description{
Serve an MCP server using stdin/stdout
}
