% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport-stdio.R
\name{StdioTransport}
\alias{StdioTransport}
\title{Stdio Transport for MCP}
\description{
Implements stdio transport for MCP communication, reading from stdin
and writing to stdout.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{server}}{The MCP server instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StdioTransport-new}{\code{StdioTransport$new()}}
\item \href{#method-StdioTransport-start}{\code{StdioTransport$start()}}
\item \href{#method-StdioTransport-stop}{\code{StdioTransport$stop()}}
\item \href{#method-StdioTransport-clone}{\code{StdioTransport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StdioTransport-new"></a>}}
\if{latex}{\out{\hypertarget{method-StdioTransport-new}{}}}
\subsection{Method \code{new()}}{
Create a new stdio transport
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StdioTransport$new(server)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{server}}{MCP server instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StdioTransport-start"></a>}}
\if{latex}{\out{\hypertarget{method-StdioTransport-start}{}}}
\subsection{Method \code{start()}}{
Start the stdio transport
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StdioTransport$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StdioTransport-stop"></a>}}
\if{latex}{\out{\hypertarget{method-StdioTransport-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the transport
Run with enhanced protocol
Run in legacy mode
Create protocol handlers for this server
Handle incoming JSON-RPC message
Handle initialize request
Handle tools/list request
Handle tools/call request
Handle resources/list request
Handle resources/read request
Handle prompts/list request
Handle prompts/get request
Create error response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StdioTransport$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StdioTransport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StdioTransport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StdioTransport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
