% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-conversion-utilities.R
\name{stream_dataframe}
\alias{stream_dataframe}
\title{Create a streaming converter for large data frames}
\usage{
stream_dataframe(df, chunk_size = 1000, callback)
}
\arguments{
\item{df}{Data frame to stream}

\item{chunk_size}{Number of rows per chunk}

\item{callback}{Function to call with each chunk}
}
\description{
Create a streaming converter for large data frames
}
