% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonrpc.R
\name{to_json}
\alias{to_json}
\title{Convert an R object to JSON}
\usage{
to_json(x, ...)
}
\arguments{
\item{x}{R object to convert to JSON}

\item{...}{Additional arguments passed to yyjsonr::write_json_str}
}
\value{
JSON string
}
\description{
Convert an R object to JSON
}
\keyword{internal}
