% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-conversion.R
\name{to_mcp_json}
\alias{to_mcp_json}
\title{Convert R objects to JSON-compatible format for MCP}
\usage{
to_mcp_json(x, auto_unbox = TRUE)
}
\arguments{
\item{x}{An R object to convert}

\item{auto_unbox}{Whether to automatically unbox single-element vectors}
}
\value{
A JSON-compatible representation of the R object
}
\description{
Convert R objects to JSON-compatible format for MCP
}
\examples{
to_mcp_json(list(a = 1, b = "hello"))
to_mcp_json(data.frame(x = 1:3, y = letters[1:3]))
}
