% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{tool_response}
\alias{tool_response}
\alias{tool_response_text}
\alias{tool_response_image}
\alias{tool_response_audio}
\alias{tool_response_resource}
\alias{tool_response_error}
\title{Create a tool_response object}
\usage{
tool_response_text(text)

tool_response_image(image, mime_type = "image/png")

tool_response_audio(audio, mime_type = "audio/mpeg")

tool_response_resource(resource)

tool_response_error(text)

tool_response(..., type = c("text", "image", "audio"), isError = FALSE)
}
\arguments{
\item{text}{Text content}

\item{image}{Image content}

\item{audio}{Audio content}

\item{resource}{Resource content}

\item{type}{Type of tool_response}

\item{isError}{Whether the tool_response is an error}
}
\value{
A tool_response object
}
\description{
Create a tool_response object
}
