% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tools_list}
\alias{tools_list}
\title{List all available tools}
\usage{
tools_list(mcp)
}
\arguments{
\item{mcp}{A server object}
}
\value{
A list containing all available tools
}
\description{
List all available tools
}
