% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-conversion.R
\name{validate_against_schema}
\alias{validate_against_schema}
\title{Validate R object against JSON schema}
\usage{
validate_against_schema(value, schema)
}
\arguments{
\item{value}{R object to validate}

\item{schema}{JSON schema definition}
}
\value{
TRUE if valid, error otherwise
}
\description{
Validate R object against JSON schema
}
