% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonrpc.R
\name{validate_request}
\alias{validate_request}
\title{Validate a JSON-RPC 2.0 request}
\usage{
validate_request(request)
}
\arguments{
\item{request}{Parsed request object}
}
\value{
NULL if valid, error response if invalid
}
\description{
Validate a JSON-RPC 2.0 request
}
\keyword{internal}
