% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-validation.R
\name{validate_resources_config}
\alias{validate_resources_config}
\title{Validate Resources Configuration}
\usage{
validate_resources_config(resources)
}
\arguments{
\item{resources}{Resources configuration}
}
\description{
Validate Resources Configuration
}
\keyword{internal}
