% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_utils.R
\name{validate_tools_config}
\alias{validate_tools_config}
\title{Validate Tools Configuration}
\usage{
validate_tools_config(tools)
}
\arguments{
\item{tools}{Tools configuration}
}
\description{
Validate Tools Configuration
}
\keyword{internal}
