% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{write}
\alias{write}
\title{Write a JSON-RPC request to a client provider}
\usage{
write(x, method, params = NULL, id = generate_id(), timeout = 5000)
}
\arguments{
\item{x}{A client provider}

\item{method}{The method to call}

\item{params}{The parameters to pass to the method}

\item{id}{The id of the request}

\item{timeout}{Timeout in milliseconds for reading the response}
}
\value{
The client provider
}
\description{
Write a JSON-RPC request to a client provider
}
