% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client}
\alias{client}
\alias{new_client_io}
\alias{new_client_http}
\alias{new_client}
\title{Create a new mcp IO}
\usage{
new_client_io(command, args = character(), name, version = "1.0.0")

new_client_http(endpoint, name, version = "1.0.0")

new_client(obj, name, version, type = c("io", "http"))
}
\arguments{
\item{command}{The command to run}

\item{args}{Arguments to pass to the command}

\item{name}{The name of the client}

\item{version}{The version of the client}

\item{endpoint}{The HTTP endpoint URL of the MCP server to connect to}

\item{obj}{The underlying object for communication (process or request)}

\item{type}{The type of client to create ("io" or "http")}
}
\value{
A new mcp client

A new MCP HTTP client object that can communicate
with the MCP server at the specified endpoint

A new mcp client object
}
\description{
Creates an MCP client that connects to an HTTP endpoint using httr2.
This client can be used to communicate with MCP servers exposed via HTTP.
}
