% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client}
\alias{client}
\alias{new_client_io}
\alias{new_client_http}
\title{Create a new mcp IO}
\usage{
new_client_io(command, args = character(), name, version = "1.0.0")

new_client_http(endpoint, name, version = "1.0.0")
}
\arguments{
\item{command}{The command to run}

\item{args}{Arguments to pass to the command}

\item{name}{The name of the client}

\item{version}{The version of the client}

\item{endpoint}{The endpoint to connect to}
}
\value{
A new mcp client
}
\description{
Create a new mcp IO
}
