% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer.R
\name{create_ellmer_type}
\alias{create_ellmer_type}
\title{Create an ellmer type function for a specific MCP property}
\usage{
create_ellmer_type(prop)
}
\arguments{
\item{prop}{The MCP property}
}
\value{
An ellmer type function call
}
\description{
Create an ellmer type function for a specific MCP property
}
\keyword{internal}
