% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonrpc.R
\name{create_response}
\alias{create_response}
\title{Create a JSON-RPC 2.0 success response}
\usage{
create_response(id, result)
}
\arguments{
\item{id}{Request identifier}

\item{result}{Result data}
}
\value{
A structured JSON-RPC 2.0 success response
}
\description{
Create a JSON-RPC 2.0 success response
}
\keyword{internal}
