% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{new_mcp}
\alias{new_mcp}
\title{Create a new MCP object}
\usage{
new_mcp(
  name,
  description,
  version,
  tools = list(),
  resources = list(),
  prompts = list()
)
}
\arguments{
\item{...}{Named arguments to be included in the MCP object}
}
\value{
A new MCP object
}
\description{
Create a new MCP object
}
\examples{
mcp <- new_mcp(name = "My MCP", description = "This is a description")
}
