% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{new_server}
\alias{new_server}
\alias{new_mcp}
\title{Create a new MCP object}
\usage{
new_server(
  name,
  description,
  version,
  tools = list(),
  resources = list(),
  prompts = list()
)

new_mcp(...)
}
\arguments{
\item{name}{Name of the MCP}

\item{description}{Description of the MCP}

\item{version}{Version of the MCP}

\item{tools}{List of tools}

\item{resources}{List of resources}

\item{prompts}{List of prompts}

\item{...}{Forwarded to \code{\link[=new_server]{new_server()}}}
}
\value{
A new MCP object
}
\description{
Create a new MCP object
}
\examples{
mcp <- new_server(
  name = "My MCP",
  description = "This is a description",
  version = "1.0.0"
)
}
