% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\alias{response_text}
\alias{response_image}
\alias{response_audio}
\alias{response_video}
\alias{response_file}
\alias{response_resource}
\alias{response_error}
\alias{response_item}
\title{Create a response object}
\usage{
response_text(text)

response_image(image, mime_type = "image/png")

response_audio(audio, mime_type = "audio/mpeg")

response_video(video, mime_type = "video/mp4")

response_file(file, mime_type = "application/octet-stream")

response_resource(resource)

response_error(text)

response_item(
  ...,
  type = c("text", "image", "audio", "video", "file", "resource")
)

response(..., is_error = FALSE)
}
\arguments{
\item{text}{Text content}

\item{image}{Image content}

\item{mime_type}{Mime type of the content}

\item{audio}{Audio content}

\item{video}{Video content}

\item{file}{File content}

\item{resource}{Resource content}

\item{...}{Mutliple \code{response} objects}

\item{type}{Type of the content}

\item{is_error}{Whether the response is an error}
}
\value{
A response object
}
\description{
Create a response object
}
\details{
Use \code{response_item} to create a custom response item.
}
\examples{
response(
  response_text("Hello, world!"),
  response_image(system.file("extdata/logo.png", package = "mcpr")),
  response_audio(system.file("extdata/sound.mp3", package = "mcpr")),
  response_video(system.file("extdata/video.mp4", package = "mcpr")),
  response_file(system.file("extdata/file.txt", package = "mcpr")),
  response_resource(system.file("extdata/resource.json", package = "mcpr"))
)

}
