% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{serve_std_io}
\alias{serve_std_io}
\title{Serve an MCP server using stdin/stdout}
\usage{
serve_std_io(mcp)
}
\arguments{
\item{mcp}{An MCP server object}
}
\value{
Nothing, runs indefinitely
}
\description{
Serve an MCP server using stdin/stdout
}
